+++ src/pysdmx/io/_pd_utils.py
