+++ src/pysdmx/io/csv/sdmx10/writer/__init__.py
@@ -6,6 +6,7 @@ from typing import Literal, Optional, Sequence, Union
 
 import pandas as pd
 
+from pysdmx.io._pd_utils import _fill_na_values
 from pysdmx.io.csv.__csv_aux_writer import (
     __write_time_period,
     _validate_schema_exists,
@@ -13,7 +14,6 @@ from pysdmx.io.csv.__csv_aux_writer import (
 from pysdmx.io.pd import PandasDataset
 from pysdmx.model import Schema
 from pysdmx.toolkit.pd._data_utils import format_labels
-from pysdmx.util._pd_utils import _fill_na_values
 
 
 def write(
