+++ src/pysdmx/io/xml/__write_structure_specific_aux.py
@@ -5,6 +5,7 @@ from typing import Any, Dict, Hashable, List
 
 import pandas as pd
 
+from pysdmx.io._pd_utils import _fill_na_values
 from pysdmx.io.pd import PandasDataset
 from pysdmx.io.xml.__write_aux import (
     ABBR_MSG,
@@ -18,7 +19,6 @@ from pysdmx.io.xml.__write_data_aux import (
 from pysdmx.io.xml.config import CHUNKSIZE
 from pysdmx.toolkit.pd._data_utils import get_codes
 from pysdmx.util import parse_short_urn
-from pysdmx.util._pd_utils import _fill_na_values
 
 
 def __memory_optimization_writing(
